requirements.txt
# KonvertIt - Production Dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.36
alembic>=1.14.0
asyncpg>=0.30.0
redis>=5.2.0
arq>=0.26.1
playwright>=1.49.0
beautifulsoup4>=4.12.3
lxml>=5.3.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
httpx>=0.28.0
bcrypt>=4.2.0
python-jose[cryptography]>=3.3.0
cryptography>=44.0.0
python-multipart>=0.0.18
structlog>=24.4.0
sentry-sdk[fastapi]>=2.19.0
gunicorn>=22.0.0

# Development Dependencies
pytest>=8.3.0
pytest-asyncio>=0.25.0
pytest-cov>=6.0.0
ruff>=0.8.0
bandit>=1.8.0
